requirements.txt
aiodns==3.2.0
aioping==0.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bokeh==3.6.3
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fastapi==0.115.11
Flask==3.0.3
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kiwisolver==1.4.8
mac_imessage==0.3.0
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.29.1
nest-asyncio==1.6.0
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.1.0
plotly==6.0.0
pluggy==1.5.0
protobuf==5.29.3
pyarrow==19.0.1
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
PyNaCl==1.5.0
pyodbc==5.2.0
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.23.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
squarify==0.4.4
starlette==0.46.1
streamlit==1.43.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.0.6
wordcloud==1.9.4
wsproto==1.2.0
xyzservices==2025.1.0
zipp==3.21.0
